Pull `Spanned` & `SpanTable` into `toc_hir`
Thins out `toc_span` and is only used by the HIR as of right now
